Proyectos/Test_Triduum-main/requirements.txt
#Django
django==4.1.7

#Django REST Framework
djangorestframework==3.14.0

#django simple history
django_simple_filter==3.2.0

#django filter
django_filter==22.1